+++ src/view/cmdbar.rs
@@ -30,6 +30,7 @@ const SELECT: KeyPair = ("S", "Select");
 const EDIT: KeyPair = ("E", "Edit");
 const MOVE: KeyPair = ("M", "Move");
 const RAISE: KeyPair = ("R", "Raise");
+const FLATTEN: KeyPair = ("F", "Flatten");
 const DOWN: KeyPair = ("J │ ↓", "Down");
 const UP: KeyPair = ("K │ ↑", "Up");
 const PROMOTE: KeyPair = ("H │ ←", "Promote");
@@ -108,7 +109,7 @@ pub fn command_bar(model: &Model) -> Line {
         Mode::Normal => normal_mode_commands(&model.state.root),
         Mode::Input(input_state) => input_mode_commands(input_state),
         Mode::Select(_) => select_mode_commands(model.state.root.size()),
-        Mode::Selected(_) => vec![EDIT, MOVE, RAISE, DELETE, CANCEL],
+        Mode::Selected(_) => vec![EDIT, MOVE, RAISE, FLATTEN, DELETE, CANCEL],
         Mode::Move(_) => vec![DOWN, UP, PROMOTE, DEMOTE, DONE, CANCEL],
         Mode::Save(_) => vec![TOGGLE, CONFIRM, CANCEL],
     };